Fix hero button visibility and add development docs
Finally fixed that annoying 'Join Our Artists' button that was basically invisible!
Changed the styling to use a dark semi-transparent background with solid white
borders so you can actually see the text now. Also added a ring effect for
extra visibility.

Also created comprehensive documentation of all the prompts and development
process we used to build this project. Should be helpful for future reference
or if someone else needs to work on this.